requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==19.10b0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.3
django-environ==0.8.1
django-filter==21.1
django-paranoid-model==1.0.7
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
PARANOID==1.2.9
pathspec==0.7.0
psycopg2==2.9.3
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
